third_party/blink/web_tests/fast/forms/number/number-keyoperation.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/common-number-keyoperation.js"></script>
</head>
<body>
<script>
description('Test for keyboard operations for &lt;input type=number>');
testNumberKeyoperations('horizontal-tb', 123);
testNumberKeyoperations('vertical-lr', 1000);
testNumberKeyoperations('vertical-rl', 5);
testNumberKeyoperations('sideways-lr', 999);
testNumberKeyoperations('sideways-rl', 42);
</script>
</body>
</html>
