third_party/blink/web_tests/fast/forms/number/number-keyoperation-expected.txt

Test for keyboard operations for <input type=number>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test starts for <input type=number style="writing-mode: horizontal-tb">
Inserting "ab123cd":
PASS input.value is "123"
Press the up arrow key:
PASS input.value is "124"
Press the down arrow key:
PASS input.value is "123"
Press the down and alt arrow key, should not decrement value:
PASS input.value is "123"
Disable input element:
PASS input.value is "123"
Read-only input element:
PASS input.value is "123"
Moving caret by pressing arrow keys:
PASS input.value is "123"
PASS input.value is "123"

Test starts for <input type=number style="writing-mode: vertical-lr">
Inserting "ab123cd":
PASS input.value is "1000"
Press the up arrow key:
PASS input.value is "1001"
Press the down arrow key:
PASS input.value is "1000"
Press the down and alt arrow key, should not decrement value:
PASS input.value is "1000"
Disable input element:
PASS input.value is "1000"
Read-only input element:
PASS input.value is "1000"
Moving caret by pressing arrow keys:
PASS input.value is "1000"
PASS input.value is "1000"

Test starts for <input type=number style="writing-mode: vertical-rl">
Inserting "ab123cd":
PASS input.value is "5"
Press the up arrow key:
PASS input.value is "6"
Press the down arrow key:
PASS input.value is "5"
Press the down and alt arrow key, should not decrement value:
PASS input.value is "5"
Disable input element:
PASS input.value is "5"
Read-only input element:
PASS input.value is "5"
Moving caret by pressing arrow keys:
PASS input.value is "5"
PASS input.value is "5"

Test starts for <input type=number style="writing-mode: sideways-lr">
Inserting "ab123cd":
PASS input.value is "999"
Press the up arrow key:
PASS input.value is "1000"
Press the down arrow key:
PASS input.value is "999"
Press the down and alt arrow key, should not decrement value:
PASS input.value is "999"
Disable input element:
PASS input.value is "999"
Read-only input element:
PASS input.value is "999"
Moving caret by pressing arrow keys:
PASS input.value is "999"
PASS input.value is "999"

Test starts for <input type=number style="writing-mode: sideways-rl">
Inserting "ab123cd":
PASS input.value is "42"
Press the up arrow key:
PASS input.value is "43"
Press the down arrow key:
PASS input.value is "42"
Press the down and alt arrow key, should not decrement value:
PASS input.value is "42"
Disable input element:
PASS input.value is "42"
Read-only input element:
PASS input.value is "42"
Moving caret by pressing arrow keys:
PASS input.value is "42"
PASS input.value is "42"
PASS successfullyParsed is true

TEST COMPLETE
